requirements.txt
flask
llama-cpp-python
faiss-cpu